ml/requirements.txt
# SmartHPA ML Module Dependencies
prometheus-api-client>=0.5.3
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
pyyaml>=6.0
python-dateutil>=2.8.0

